KOGITO-7766 - Data Index as Quarkus extension
