+++ data-index/kogito-addons-quarkus-data-index-common/runtime/src/main/java/org/kie/kogito/index/graphql/GraphQLSchemaManagerImpl.java
