--- data-index/kogito-addons-quarkus-data-index-inmemory/integration-tests/src/main/java/org/kie/kogito/addons/quarkus/data/index/it/KogitoAddonsQuarkusDataIndexResource.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2022 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kie.kogito.addons.quarkus.data.index.it;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-
-@Path("/kogito-addons-quarkus-data-index")
-@ApplicationScoped
-public class KogitoAddonsQuarkusDataIndexResource {
-
-    @GET
-    public String hello() {
-        return "Hello kogito-addons-quarkus-data-index";
-    }
-}
