--- data-index/kogito-addons-quarkus-data-index-inmemory/integration-tests/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-quarkus.devservices.enabled=false
