--- data-index/kogito-addons-quarkus-data-index-inmemory/integration-tests/src/test/java/org/kie/kogito/addons/quarkus/data/index/it/KogitoAddonsQuarkusDataIndexResourceTest.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2022 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.kogito.addons.quarkus.data.index.it;
-
-import org.junit.jupiter.api.Test;
-
-import io.quarkus.test.junit.QuarkusTest;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.Matchers.is;
-
-@QuarkusTest
-public class KogitoAddonsQuarkusDataIndexResourceTest {
-
-    @Test
-    public void testHelloEndpoint() {
-        given()
-                .when().get("/kogito-addons-quarkus-data-index")
-                .then()
-                .statusCode(200)
-                .body(is("Hello kogito-addons-quarkus-data-index"));
-    }
-}
