+++ data-index/kogito-addons-quarkus-data-index-inmemory/runtime/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 #Data Index
-kogito.persistence.type=postgresql
+kogito.apps.persistence.type=postgresql
 kogito.data-index.domain-indexing=false
 kogito.data-index.blocking=true
