+++ data-index/kogito-addons-quarkus-data-index-postgresql/deployment/pom.xml
@@ -4,11 +4,11 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.kie.kogito</groupId>
-    <artifactId>kogito-addons-quarkus-data-index-inmemory-parent</artifactId>
+    <artifactId>kogito-addons-quarkus-data-index-postgresql-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
-  <artifactId>kogito-addons-quarkus-data-index-inmemory-deployment</artifactId>
-  <name>Kogito Apps :: Kogito Addons Quarkus Data Index In-memory - Deployment</name>
+  <artifactId>kogito-addons-quarkus-data-index-postgresql-deployment</artifactId>
+  <name>Kogito Apps :: Kogito Addons Quarkus Data Index PostgreSQL - Deployment</name>
   <dependencies>
     <dependency>
       <groupId>org.kie.kogito</groupId>
@@ -16,11 +16,7 @@
     </dependency>
     <dependency>
       <groupId>org.kie.kogito</groupId>
-      <artifactId>kogito-addons-quarkus-data-index-inmemory</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.kie.kogito</groupId>
-      <artifactId>quarkus-postgresql-inmemory-deployment</artifactId>
+      <artifactId>kogito-addons-quarkus-data-index-postgresql</artifactId>
     </dependency>
     <dependency>
       <groupId>io.quarkus</groupId>
