+++ data-index/kogito-addons-quarkus-data-index-postgresql/integration-tests/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+kogito.persistence.type=jdbc
+kogito.persistence.auto.ddl=true
+
+quarkus.datasource.db-kind=postgresql
