+++ data-index/kogito-addons-quarkus-data-index-postgresql/integration-tests/src/main/resources/greet.sw.json
@@ -1,0 +1,71 @@
+{
+  "id": "greet",
+  "version": "1.0",
+  "name": "Greeting workflow",
+  "expressionLang": "jsonpath",
+  "description": "JSON based greeting workflow",
+  "start": "ChooseOnLanguage",
+  "functions": [
+    {
+      "name": "greetFunction",
+      "type": "custom",
+      "operation": "sysout"
+    },
+    {
+       "name": "isEnglish",
+       "type": "expression",
+       "operation" : "$.[?(@.language  == 'English')]" 
+    }
+  ],
+  "states": [
+    {
+      "name": "ChooseOnLanguage",
+      "type": "switch",
+      "dataConditions": [
+        {
+          "condition": "fn:isEnglish",
+          "transition": "GreetInEnglish"
+        },
+        {
+          "condition": "${ $.[?(@.language  == 'Spanish')] }",
+          "transition": "GreetInSpanish"
+        }
+      ],
+      "defaultCondition": {
+        "transition": "GreetInEnglish"
+      }
+    },
+    {
+      "name": "GreetInEnglish",
+      "type": "inject",
+      "data": {
+        "greeting": "Hello from JSON Workflow,"
+      },
+      "transition": "GreetPerson"
+    },
+    {
+      "name": "GreetInSpanish",
+      "type": "inject",
+      "data": {
+        "greeting": "Saludos desde JSON Workflow,"
+      },
+      "transition": "GreetPerson"
+    },
+    {
+      "name": "GreetPerson",
+      "type": "operation",
+      "actions": [
+        {
+          "name": "greetAction",
+          "functionRef": {
+            "refName": "greetFunction",
+            "arguments": {
+              "message": "$.greeting $.name"
+            }
+          }
+        }
+      ],
+      "end": "true"
+    }
+  ]
+}
