+++ data-index/kogito-addons-quarkus-data-index-postgresql/integration-tests/src/test/java/org/kie/kogito/addons/quarkus/data/index/it/KogitoAddonsQuarkusDataIndexResourceIT.java
