+++ data-index/kogito-addons-quarkus-data-index-postgresql/pom.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.kie.kogito</groupId>
+    <artifactId>data-index</artifactId>
+    <version>2.0.0-SNAPSHOT</version>
+  </parent>
+  <artifactId>kogito-addons-quarkus-data-index-postgresql-parent</artifactId>
+  <packaging>pom</packaging>
+
+  <name>Kogito Apps :: Kogito Addons Quarkus Data Index PostgreSQL - Parent</name>
+
+  <modules>
+    <module>deployment</module>
+    <module>runtime</module>
+    <module>integration-tests</module>
+  </modules>
+
+</project>
