+++ data-index/kogito-addons-quarkus-data-index-postgresql/runtime/pom.xml
@@ -4,11 +4,11 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.kie.kogito</groupId>
-    <artifactId>kogito-addons-quarkus-data-index-inmemory-parent</artifactId>
+    <artifactId>kogito-addons-quarkus-data-index-postgresql-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
-  <artifactId>kogito-addons-quarkus-data-index-inmemory</artifactId>
-  <name>Kogito Apps :: Kogito Addons Quarkus Data Index In-memory - Runtime</name>
+  <artifactId>kogito-addons-quarkus-data-index-postgresql</artifactId>
+  <name>Kogito Apps :: Kogito Addons Quarkus Data Index PostgreSQL - Runtime</name>
   <dependencies>
     <dependency>
       <groupId>org.kie.kogito</groupId>
@@ -17,10 +17,6 @@
     <dependency>
       <groupId>org.kie.kogito</groupId>
       <artifactId>data-index-storage-postgresql</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.kie.kogito</groupId>
-      <artifactId>quarkus-postgresql-inmemory</artifactId>
     </dependency>
     <dependency>
       <groupId>io.quarkus</groupId>
