+++ data-index/kogito-addons-quarkus-data-index-postgresql/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,9 @@
+name: Kogito Addons Quarkus Data Index PostgreSQL
+description: Kogito Data Index PostgreSQL Add-On
+metadata:
+  keywords:
+    - data-index
+    - kogito
+  guide: https://quarkus.io/guides/kogito
+  categories:
+    - "business-automation"
