+++ data-index/kogito-addons-quarkus-data-index-postgresql/runtime/src/main/resources/application.properties
@@ -1,0 +1,28 @@
+#
+# Copyright 2022 Red Hat, Inc. and/or its affiliates.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+#Data Index
+kogito.apps.persistence.type=postgresql
+kogito.data-index.domain-indexing=false
+kogito.data-index.blocking=true
+
+#PostgreSql
+quarkus.datasource.db-kind=postgresql
+
+#Hibernate
+quarkus.hibernate-orm.jdbc.timezone=UTC
+quarkus.hibernate-orm.database.generation=update
+quarkus.hibernate-orm.database.generation.halt-on-error=true
