+++ kogito-apps-bom/pom.xml
@@ -233,6 +233,28 @@
                 <version>${project.version}</version>
                 <classifier>sources</classifier>
             </dependency>
+            <dependency>
+              <groupId>org.kie.kogito</groupId>
+              <artifactId>kogito-addons-quarkus-data-index-common-runtime</artifactId>
+              <version>${project.version}</version>
+            </dependency>
+            <dependency>
+              <groupId>org.kie.kogito</groupId>
+              <artifactId>kogito-addons-quarkus-data-index-common-runtime</artifactId>
+              <version>${project.version}</version>
+              <classifier>sources</classifier>
+            </dependency>
+            <dependency>
+              <groupId>org.kie.kogito</groupId>
+              <artifactId>kogito-addons-quarkus-data-index-common-deployment</artifactId>
+              <version>${project.version}</version>
+            </dependency>
+            <dependency>
+              <groupId>org.kie.kogito</groupId>
+              <artifactId>kogito-addons-quarkus-data-index-common-deployment</artifactId>
+              <version>${project.version}</version>
+              <classifier>sources</classifier>
+            </dependency>
 
 
             <!-- Kogito Apps Persistence -->
